backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-groq>=0.1.0
spacy>=3.7.4
pydantic>=2.6.1
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic
